// src/pages/Profile/types.ts
export interface Trip {
  id: number;
  title: string;
  image: string;
  timeAgo: string;
}

export interface UserProfile {
  fullName: string;
  email: string;
  phone: string;
  address: string;
  bio: string;
  badge: string;
  avatarUrl: string;
  coverUrl: string;
  joinDate: string;
  location: string;
}